rename onboarding related files
